requirements.txt
jmespath==0.10.0
picamera==1.13
python-dateutil==2.8.1
six==1.15.0
urllib3==1.26.2
requests==2.25.1
Flask==1.1.2